+++ include/seqan3/core/simd/detail/simd_algorithm_sse4.hpp
@@ -16,7 +16,7 @@
 
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/simd/detail/builtin_simd_intrinsics.hpp>
-#include <seqan3/core/simd/detail/builtin_simd.hpp>
+#include <seqan3/utility/simd/detail/builtin_simd.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
 
 //-----------------------------------------------------------------------------
