+++ test/unit/core/simd/detail/default_simd_backend_test.cpp
@@ -8,7 +8,7 @@
 #include <gtest/gtest.h>
 
 #include <seqan3/utility/simd/concept.hpp>
-#include <seqan3/core/simd/detail/builtin_simd.hpp>
+#include <seqan3/utility/simd/detail/builtin_simd.hpp>
 #include <seqan3/core/simd/detail/default_simd_backend.hpp>
 
 #include <iostream>
