+++ test/unit/utility/simd/detail/builtin_simd_test.cpp
@@ -11,7 +11,7 @@
 
 #include <seqan3/core/detail/pack_algorithm.hpp>
 #include <seqan3/utility/simd/concept.hpp>
-#include <seqan3/core/simd/detail/builtin_simd.hpp>
+#include <seqan3/utility/simd/detail/builtin_simd.hpp>
 #include <seqan3/utility/type_list/type_list.hpp>
 
 #include <iostream>
